Vitamin-D
Vitamin D is an essential vitamin to have running through your system. It helps to maintain the essential functions of your body. Vitamin D maintains the health of your bones and teeth, supports the immune system, supports peoples brain function, supports the nervous system, regulates insulin and diabetes management, supports lung function, helps with cardiovascular health, and can influence the expression of genes involved in cancer development. Vitamin D is an important part of a well functioning body. Vitamin D is actually considered a prohormone and not a vitamin. It can be synthesized by our bodies through sunlight. Vitamin D cannot be produced by our bodies and must be taken in through diet or sunlight. The recommended amount of sunlight to get in order to fulfill your bodies requirement of Vitamin D is 5-10 minutes about 3-4 times per week. Your skin will take in the sunlight and your body can naturally produce vitamin D from the sun's rays. It is important to get this essential sunlight to get the proper amount of Vitamin D in your system. Especially in the winter months, it can be difficult to get a good amount of sunlight. Without Vitamin D you can see some serious health effects that can impact your body. If you are somewhere in the far north,  where during winter you do not see much sun, it is essential to substitute different foods or take supplements in order to get the proper amount of Vitamin D. 

You can also get Vitamin D from certain foods if you are not getting the proper amount of sunlight. These foods include fish, fish liver oils, egg yolks, fortified dairy and grain products. These are great foods to eat if you are not getting the proper amount of Vitamin D. It can be difficult for vegans especially to maintain the proper amount of Vitamin D as they do not eat foods that will contain the vitamin. Many vegans will have to take supplements in order to meet the proper amount of Vitamin D so as not to have a deficiency. Vitamin D deficiency can cause some serious health effects. These include: 
-	Cognitive impairment 
-	Increased risk of cardiovascular disease 
-	Severe asthma 
-	Cancer 
The lack of the vitamin also may leave people more at risk for diabetes, rickets, soft bones, and skeletal deformation. It is extremely important to maintain the proper amount of Vitamin D within your system. Whether this is by getting the proper amount of sunlight, eating certain foods or taking supplements. It is important to maintain all the proper vitamins within your system so that your body con function properly. Without the proper amount of vitamins in your system, it can lead your body to have serious health defects. In the future, it is a good idea to keep in mind Vitamin D, especially in the winter months. It can get very cold and unpleasant outdoors making it difficult to get the proper amount of sunlight. In order to combat this, you must en corporate certain foods in your diet or take Vitamin D supplements so you do not become deficient. 

Amber Whipple